Crate splendor_core

Core game structures implementation.

Structs

ActionRecord

Wraps an action with the player who took it.

ActionRequest

A struct to represent an action request.

ActorError

An error type for actor errors.

BuyCardAction

A struct to represent the buy card action.

Card

A struct to represent a card.

CardPoolSnapshot

A struct to represent the card pool snapshot.

ColorIter

An iterator over the variants of Color

ColorVec

A struct to represent the color combinations.

DevelopmentCards

A struct to represent the development cards in player’s hand.

DropTokensAction

A struct to represent the drop tokens action.

GameSnapshot

A struct to represent the game snapshot.

Noble

A struct to represent a noble.

PlayerSnapshot

A struct to represent the player snapshot.

ReserveCardAction

A struct to represent the reserve card action.

ReservedCard

A struct to represent a reserved card.

SelectNoblesAction

A struct to represent the select nobles action.

TierIter

An iterator over the variants of Tier

Enums

ActionType

An enum to represent the actions a player can take.

BuyCardSource

An enum to represent the source of the card to buy.

CardView

A struct to represent the view of other players’ reserved cards.

Color

An enum to represent the colors.

PlayerAction

An enum to represent the actions a player can take.

Record

A record of a game event.

TakeTokenAction

An enum to represent the take tokens action.

Tier

An enum to represent the card tiers.

Constants

MAX_PLAYERS

The maximum number of players in a game.

Traits

PlayerActor

A player actor trait.