Struct splendor_core::ColorVec

pub struct ColorVec(/* private fields */);

A struct to represent the color combinations.

Note

when comparing two ColorVecs a, b, the order is defined as:

a < b iff \forall i, a[i] <= b[i] and \exists j, a[j] < b[j]

a <= b iff \forall i, a[i] <= b[i]

a > b iff \forall i, a[i] >= b[i] and \exists j, a[j] > b[j]

a >= b iff \forall i, a[i] >= b[i]

otherwise a and b are not comparable

Also the equality is trivially defined as the equality of the vectors.

Hence, the color vector cannot impl PartialOrd and Ord, following stmts are false:

a <= b iff a < b || a == b

a >= b iff a > b || a == b

Implementations

impl ColorVec

pub const fn empty() -> Self

Create a new empty color vector.

pub const fn new( black: u8, blue: u8, green: u8, red: u8, white: u8, yellow: u8 ) -> Self

Create a new color vector with the given values.

pub fn get(&self, color: Color) -> u8

Get the value of a color.

pub fn set(&mut self, color: Color, value: u8)

Set the value of a color.

pub fn add(&mut self, color: Color, value: u8)

Add a value to a color.

pub fn sub(&mut self, color: Color, value: u8)

Sub a value to a color.

pub fn iter(&self) -> impl Iterator<Item = u8> + '_

Get an iterator over the colors.

pub fn total(&self) -> u8

Get the total number of tokens.

pub fn saturating_sub(&self, rhs: &Self) -> Self

Saturation subtraction.

pub fn lt(&self, rhs: &Self) -> bool

Check if the color vector is less than the other.

The order is defined as: \forall i, a[i] <= b[i] and \exists j, a[j] < b[j]

pub fn le(&self, rhs: &Self) -> bool

Check if the color vector is less than or equal to the other.

The order is defined as: \forall i, a[i] <= b[i]

pub fn gt(&self, rhs: &Self) -> bool

Check if the color vector is greater than the other.

The order is defined as: \forall i, a[i] >= b[i] and \exists j, a[j] > b[j]

pub fn ge(&self, rhs: &Self) -> bool

Check if the color vector is greater than or equal to the other.